+++ labkey_multisite_query_tool/__init__.py
@@ -3,3 +3,5 @@
 __author__ = 'Stefan Novak'
 __email__ = 'novast@ohsu.edu'
 __version__ = '0.1.0'
+
+from .labkey import *